+++ MammalAPI/Controllers/HabitatController.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using MammalAPI.HATEOAS;
+//using MammalAPI.HATEOAS;
